Use plugin:// stream for resolving sandbox images
I was having issues with the sandbox images displaying on my
development environment, and found it to be because the images couldn't
be loaded. This fix changes how the URL is constructed to allow finding
the appropriate sandbox image based on the plugin:// URL
stream/protocol instead of constructing it via the base_url_absolute
variable.

Using the plugin:// protocol appears to be a little more portable.